feat: class 분리 및 getGuessLotto 함수 추가
class Lotto에서 더 적은 기능을 수행하도록 class를 재분리하였고 사용자로부터 로또 번호를 입력받아오는 getGuessLotto 함수를 추가하였습니다.
