+++ docs/README.md
@@ -1,4 +1,8 @@
 # 기능 목록
+### 클래스 나누기
+- App.js, Lotto.js의 클래스를 나누기
+- Lotto.js는 하나의 로또의 통계를 계산하는 기능
+- App.js는 Lotto.js의 기능을 제외한 모든 기능
 ### 로또 구입금액 입력받기 ⭕
 - 입력을 요구하는 문자열 출력하기 ⭕
 - 1,000원 단위의 숫자로 입력받기 ⭕
