+++ src/Constants.js
@@ -1,6 +1,6 @@
 const GAME_MESSAGE = {
-  parchaseAmount: '구입금액을 입력해 주세요.\n',
-  guessNumbers: '당첨 번호를 입력해 주세요.',
+  purchaseAmount: '구입금액을 입력해 주세요.\n',
+  guessNumbers: '\n당첨 번호를 입력해 주세요.\n',
   bonusNumbers: '보너스 번호를 입력해 주세요.',
   statics: '당첨 통계\n---',
 };
