+++ src/Lotto.js
@@ -1,5 +1,5 @@
 const { Console, Random } = require('@woowacourse/mission-utils');
-const { validateMoney } = require('./InputChecker');
+
 
 class Lotto {
   #numbers
@@ -9,7 +9,6 @@ class Lotto {
     this.#numbers = numbers;
   }
 
-  // TypeError: Cannot read property 'length' of undefined
   validate(numbers) {
     if (numbers.length !== 6) {
       throw new Error("[ERROR] 로또 번호는 6개여야 합니다.");
@@ -17,33 +16,12 @@ class Lotto {
   }
 
   // TODO: 추가 기능 구현
-  getParchaseAmount(message) {
-    Console.readLine(message, (inputMoney) => {
-      validateMoney(inputMoney);
-      this.printNumberOfLottos(inputMoney);
-    })
+  validateLotto() {
+    this.validate(this.#numbers);
+    Console.print('당첨 번호를 받는 것까지 성공했습니다.');
   }
 
-  printNumberOfLottos(inputMoney) {
-    const numberOfLotto = Math.floor(inputMoney / 1000);
-    Console.print('\n' + numberOfLotto + '개를 구매했습니다.');
-    this.printLottos(numberOfLotto);
-  }
-
-  printLottos(numberOfLotto) {
-    const lottoArrays = [];
-    for (let i = 0; i < numberOfLotto; i++) {
-      this.#numbers = Random.pickUniqueNumbersInRange(1, 45, 6);
-      console.log(this.#numbers);
-      this.validate(this.#numbers);
-      console.log(this.#numbers);
-      lottoArrays.push(this.#numbers);
-    }
-    lottoArrays.forEach((numbers) => {
-      return Console.print(numbers);
-      } 
-    )
-  }
+  
 }
 
 module.exports = Lotto;
